+++ changelog.md
@@ -22,7 +22,9 @@ All notable changes to this project will be documented in this file.
 - Implement the following types in `kotools.types.collections`:
     - `NotEmptyList` ([#6](https://github.com/kotools/types/issues/6))
     - `NotEmptyCollection` and `NotEmptySet`
-      ([#7](https://github.com/kotools/types/issues/7)).
+      ([#7](https://github.com/kotools/types/issues/7))
+    - `NotEmptyMutableCollection` and `NotEmptyMutableList`
+      ([#12](https://github.com/kotools/types/issues/12)).
 - Overload comparisons of the following types in `kotools.types.number`:
     - `NonZeroInt` ([#33](https://github.com/kotools/types/issues/33))
     - `PositiveInt` ([#34](https://github.com/kotools/types/issues/34))